Integrate with 'Open URL in container' extension
This allows to click on a URL and have it open in a new temp container, and let this extension track that container just as if it was created using the extension.
In order to trigger this, a new container name, as passed from 'Open URL in container', must equal to '%TEMP%'.