Handle %TEMP% at container creation
When a container is updated, if the active tab is in that container,
it does not show the updated name, color, and symbol until it ceases
to be the active tab.

Modifying the container at the time of *its* creation, before the tab
is created, avoids this (at least sometimes).